+++ main.cpp
@@ -1,4 +1,4 @@
-#include "main.h"
+#include "Doctor.h"
 #include <vector>
 
 using namespace std;
@@ -6,80 +6,8 @@ using namespace std;
 const int addPatient = 1;
 const int showPatients = 2;
 const int healPatients = 3;
-const int exit = 4;
+const int exitNum = 4;
 
-void Doctor::setName(string n)
-{
-    name = n;
-}
-
-string Doctor::getName()
-{
-    return name;
-}
-
-void Doctor::setAge(int a)
-{
-    age = a;
-}
-
-int Doctor::getAge()
-{
-    return age;
-}
-
-void Doctor::setExperience(int e)
-{
-    experience = e;
-}
-
-int Doctor::getExperience() const
-{
-    return experience;
-}
-
-void Doctor::setSpecialty(string s)
-{
-    specialty = s;
-}
-
-string Doctor::getSpecialty()
-{
-    return specialty;
-}
-
-void Doctor::setHospital(string h)
-{
-    hospital = h;
-}
-
-string Doctor::getHospital()
-{
-    return hospital;
-}
-
-void Doctor::addPatient(string p)
-{
-    patients.push_back(p);
-}
-
-void Doctor::showPatients()
-{
-    for (string p : patients)
-        cout << "\n"<< p;
-}
-
-void Doctor::healPatient(string p)
-{
-    const auto it = find(patients.begin(), patients.end(), p);
-    if (it != patients.end())
-    {
-        patients.erase(it);
-        cout << "Patient healed successfully";
-    }
-    else
-        cout << "Write correct patient name.";
-}
 
 void mainLogic(Doctor doctor)
 {
@@ -113,7 +41,7 @@ void mainLogic(Doctor doctor)
             cin >> patient;
             doctor.healPatient(patient);
             break;
-        case exit:
+        case exitNum:
             isWorking = false;
             break;
         default:
